refactor: move shared cluster components to clusters/ level
Move cluster-general and user-permissions components from
kafka-connections/details/ to clusters/ since they are shared
between kafka-clusters and kafka-connections detail pages.

Deploy/undeploy and lifecycle badge conditioned to KAFKA_CLUSTER
type only, using @if syntax.
