+++ gravitee-apim-console-webui/src/management/clusters/general/cluster-general.component.scss
@@ -2,7 +2,7 @@
 @use '@angular/material/index' as mat;
 @use '@gravitee/ui-particles-angular/index' as gio;
 
-@use '../../../../../scss/gio-layout' as gio-layout;
+@use '../../../scss/gio-layout' as gio-layout;
 
 $typography: map.get(gio.$mat-theme, typography);
