+++ gravitee-apim-console-webui/src/management/clusters/user-permissions/manage-groups-dialog/cluster-manage-groups-dialog.component.scss
@@ -1,7 +1,7 @@
 @use 'sass:map';
 @use '@angular/material' as mat;
 @use '@gravitee/ui-particles-angular' as gio;
-@use '../../../../../../scss/gio-layout' as gio-layout;
+@use '../../../../scss/gio-layout' as gio-layout;
 $typography: map.get(gio.$mat-theme, typography);
 
 .content {
