+++ gravitee-apim-console-webui/src/management/clusters/user-permissions/manage-groups-dialog/cluster-manage-groups-dialog.harness.ts
