+++ gravitee-apim-console-webui/src/management/clusters/user-permissions/transfer-ownership/cluster-transfer-ownership-dialog.component.scss
@@ -1,4 +1,4 @@
-@use '../../../../../../scss/gio-layout' as gio-layout;
+@use '../../../../scss/gio-layout' as gio-layout;
 @use '@angular/material' as mat;
 @use '@gravitee/ui-particles-angular' as gio;
 @use 'sass:map';
