+++ gravitee-apim-console-webui/src/management/clusters/user-permissions/transfer-ownership/cluster-transfer-ownership-dialog.harness.ts
